+++ src/errors/mod.rs
@@ -4,7 +4,10 @@ mod collection_collision;
 mod collection_error;
 mod item_collision;
 mod result_collection_error;
+
+#[cfg(feature = "tuple")]
 mod tuple_extension_error;
+#[cfg(feature = "tuple")]
 mod unzip_error;
 
 pub use capacity_mismatch::*;
@@ -12,5 +15,8 @@ pub use collection_collision::*;
 pub use collection_error::*;
 pub use item_collision::*;
 pub use result_collection_error::*;
+
+#[cfg(feature = "tuple")]
 pub use tuple_extension_error::*;
+#[cfg(feature = "tuple")]
 pub use unzip_error::*;