+++ src/traits/mod.rs
@@ -1,7 +1,11 @@
 mod try_extend;
 mod try_from_iterator;
+
+#[cfg(feature = "tuple")]
 mod try_unzip;
 
 pub use try_extend::*;
 pub use try_from_iterator::*;
+
+#[cfg(feature = "tuple")]
 pub use try_unzip::TryUnzip;