+++ tests/doc/main.rs
@@ -1,6 +1,0 @@
-mod array;
-mod arrayvec;
-mod result;
-mod try_extend;
-mod try_unzip;
-mod tuples;